SVN r16284 |2011-05-20 19:38:10 +0000

+++ ehr/resources/queries/ehr/cage_observations.js
@@ -9,80 +9,84 @@ var {EHR, LABKEY, Ext, console, init, beforeInsert, afterInsert, beforeUpdate, a
 
 
 
-//TODO: cascade deletes to ehr.cage_obs
+//TODO : cascade deletes to ehr.cage_obs
 
 
 //insert into ehr.cage_obs
-function onAfterInsert(errors, scriptContext, row, oldRow){
-    //find animals overlapping with this record
-    if(row.Date && row.room){
-        var sql = "SELECT * FROM study.housing h " +
-        "WHERE h.room='"+row.room+"' AND " +
-        "h.date <= '"+row.Date+"' AND " +
-        "(h.enddate >= '"+row.Date+"' OR h.enddate IS NULL) AND " +
-        "h.qcstate.publicdata = true ";
+function onAfterInsert(scriptContext, errors, row, oldRow){
+    if(scriptContext.extraContext.dataSource != 'etl'){
+        //find animals overlapping with this record
+        if(row.Date && row.room){
+            var sql = "SELECT Id, room, cage, lsid FROM study.housing h " +
+            "WHERE h.room='"+row.room+"' AND " +
+            "h.date <= '"+row.Date+"' AND " +
+            "(h.enddate >= '"+row.Date+"' OR h.enddate IS NULL) AND " +
+            "h.qcstate.publicdata = true ";
 
-        if(row.cage)
-            sql += " AND h.cage='"+row.cage+"'";
-        console.log(sql);
-        var toInsert = [];
-        LABKEY.Query.executeSql({
-            schemaName: 'study',
-            sql: sql,
-            success: function(data){
-                if(data && data.rows && data.rows.length){
-                    var obj;
-                    Ext.each(data.rows, function(r){
-                        obj = {Id: r.Id, QCStateLabel: row.QCStateLabel, RoomAtTime: r.room, date: new Date(row.date), observationRecord: row.lsid, housingRecord: r.lsid, remark: row.remark, taskid: row.taskid, performedby: row.performedby};
-
-                        if(row.cage)
-                            obj.CageAtTime = r.cage;
-
-                        toInsert.push(obj);
-                    }, this);
-                }
-                else {
-                    console.log('No animals found in this room/cage: '+row.room + '/'+row.cage)
-                }
-            },
-            failure: EHR.onFailure
-        });
-
-        if(toInsert.length){
-            LABKEY.Query.insertRows({
+            if(row.cage)
+                sql += " AND h.cage='"+row.cage+"'";
+            //console.log(sql);
+            var toInsert = [];
+            LABKEY.Query.executeSql({
                 schemaName: 'study',
-                queryName: 'Cage Observations',
-                rows: toInsert,
-                scope: this,
+                sql: sql,
                 success: function(data){
-                    console.log('Success Cascade Inserting')
+                    if(data && data.rows && data.rows.length){
+                        var obj;
+                        Ext.each(data.rows, function(r){
+                            obj = {Id: r.Id, QCStateLabel: row.QCStateLabel, RoomAtTime: r.room, date: new Date(row.date), observationRecord: row.objectid, housingRecord: r.lsid, remark: row.remark, taskid: row.taskid, performedby: row.performedby};
+
+                            if(row.cage)
+                                obj.CageAtTime = r.cage;
+
+                            toInsert.push(obj);
+                        }, this);
+                    }
+                    else {
+                        console.log('No animals found in this room/cage: '+row.room + '/'+row.cage)
+                    }
                 },
-                failure: function(error){
-                    console.log(error.exceptionClass)
-                }
+                failure: EHR.onFailure
             });
+
+            if(toInsert.length){
+                LABKEY.Query.insertRows({
+                    schemaName: 'study',
+                    queryName: 'Cage Observations',
+                    rows: toInsert,
+                    scope: this,
+                    success: function(data){
+                        console.log('Success Cascade Inserting')
+                    },
+                    failure: function(error){
+                        console.log(error.exceptionClass)
+                    }
+                });
+            }
         }
     }
 }
 
 //TODO
-//insert into study.cage_obs
+
 //function onAfterUpdate(errors, scriptContext, row, oldRow){
 //    //find animals overlapping with this record
 //    if(row.Date && row.room){
 //        //first we locate existing records to avoid unnecessary inserts
 //        var existingRecords = [];
 //        var distinctIds = [];
+//        var distinctObservations = [];
 //        LABKEY.Query.selectRows({
 //            schemaName: 'study',
 //            queryName: 'Cage Observations',
 //            filterArray: [
-//                LABKEY.Filter.create('observationRecord', row.lsid, LABKEY.Filter.Types.EQUAL)
+//                LABKEY.Filter.create('observationRecord', row.objectid, LABKEY.Filter.Types.EQUAL)
 //            ],
 //            scope: this,
 //            success: function(data){
 //                Ext.each(data.rows, function(r){
 //                    distinctIds.push(r.Id);
+//                    distinctObservations.push(r.observationRecord);
 //                    existingRecords.push(r);
 //                }, this);
 //            },
@@ -90,7 +94,8 @@ function onAfterInsert(errors, scriptContext, row, oldRow){
 //        });
 //        distinctIds = distinctIds.unique();
 //
-//        var sql = "SELECT * FROM study.housing h " +
+//        //then we find records that should exist
+//        var sql = "SELECT Id, room, cage, lsid FROM study.housing h " +
 //        "WHERE h.room='"+row.room+"' AND " +
 //        "h.date <= '"+row.Date+"' AND " +
 //        "(h.enddate >= '"+row.Date+"' OR h.enddate IS NULL) AND " +
@@ -100,6 +105,8 @@ function onAfterInsert(errors, scriptContext, row, oldRow){
 //            sql += " AND h.cage='"+row.cage+"'";
 //        //console.log(sql);
 //        var toInsert = [];
+//        var toUpdate = [];
+//        var toDelete = [];
 //        LABKEY.Query.executeSql({
 //            schemaName: 'study',
 //            sql: sql,
@@ -107,12 +114,19 @@ function onAfterInsert(errors, scriptContext, row, oldRow){
 //                if(data && data.rows && data.rows.length){
 //                    var obj;
 //                    Ext.each(data.rows, function(r){
-//                        obj = {Id: r.Id, QCStateLabel: row.QCStateLabel, RoomAtTime: r.room, date: new Date(row.date), observationRecord: row.lsid, housingRecord: r.lsid, remark: row.remark, taskid: row.taskid, performedby: row.performedby};
+//                        //only insert the row if it doesnt exist
+//                        if(distinctIds.indexOf(r.Id) == -1){
+//                            obj = {Id: r.Id, QCStateLabel: row.QCStateLabel, RoomAtTime: r.room, date: new Date(row.date), observationRecord: row.lsid, housingRecord: r.lsid, remark: row.remark, taskid: row.taskid, performedby: row.performedby};
 //
-//                        if(row.cage)
-//                            obj.CageAtTime = r.cage;
+//                            if(row.cage)
+//                                obj.CageAtTime = r.cage;
 //
-//                        toInsert.push(obj);
+//                            toInsert.push(obj);
+//                        }
+//                        else if (r.remark != row.remark){
+//                            obj = {Id: r.Id, QCStateLabel: row.QCStateLabel, RoomAtTime: r.room, date: new Date(row.date), observationRecord: row.lsid, housingRecord: r.lsid, remark: row.remark, taskid: row.taskid, performedby: row.performedby};
+//                            toUpdate.push(obj);
+//                        }
 //                    }, this);
 //                }
 //                else {
@@ -140,38 +154,40 @@ function onAfterInsert(errors, scriptContext, row, oldRow){
 //}
 
 //cascade delete
-function onAfterDelete(errors, scriptContext, row, oldRow){
-    var toDelete = [];
-    if(row.lsid){
-        LABKEY.Query.selectRows({
-            schemaName: 'study',
-            queryName: 'Cage Observations',
-            filterArray: [
-                LABKEY.Filter.create('observationRecord', row.lsid, LABKEY.Filter.Types.EQUAL)
-            ],
-            scope: this,
-            success: function(data){
-                Ext.each(data.rows, function(r){
-                    toDelete.push({lsid: r.lsid});
-                }, this);
-            },
-            failure: EHR.onFailure
-        });
-    }
-console.log(toDelete);
-    if(toDelete.length){
-        LABKEY.Query.deleteRows({
-            schemaName: 'study',
-            queryName: 'Cage Observations',
-            rows: toDelete,
-            scope: this,
-            success: function(data){
-                console.log('Success Cascade Deleting');
-            },
-            failure: function(error){
-                console.log(error.exceptionClass)
-            }
-        });
+function onAfterDelete(scriptContext, errors, row, oldRow){
+    if(scriptContext.extraContext.dataSource != 'etl'){
+        var toDelete = [];
+        if(row.lsid){
+            LABKEY.Query.selectRows({
+                schemaName: 'study',
+                queryName: 'Cage Observations',
+                filterArray: [
+                    LABKEY.Filter.create('observationRecord', row.lsid, LABKEY.Filter.Types.EQUAL)
+                ],
+                scope: this,
+                success: function(data){
+                    Ext.each(data.rows, function(r){
+                        toDelete.push({lsid: r.lsid});
+                    }, this);
+                },
+                failure: EHR.onFailure
+            });
+        }
+        //console.log(toDelete);
+        if(toDelete.length){
+            LABKEY.Query.deleteRows({
+                schemaName: 'study',
+                queryName: 'Cage Observations',
+                rows: toDelete,
+                scope: this,
+                success: function(data){
+                    console.log('Success Cascade Deleting');
+                },
+                failure: function(error){
+                    console.log(error.exceptionClass)
+                }
+            });
+        }
     }
 }
